Suppress int()

+++ isatty_darwin.go
@@ -9,8 +9,8 @@ import (
 
 const ioctlReadTermios = syscall.TIOCGETA
 
-func IsTerminal(fd int) bool {
+func IsTerminal(fd uintptr) bool {
 	var termios syscall.Termios
-	_, _, err := syscall.Syscall6(syscall.SYS_IOCTL, uintptr(fd), ioctlReadTermios, uintptr(unsafe.Pointer(&termios)), 0, 0, 0)
+	_, _, err := syscall.Syscall6(syscall.SYS_IOCTL, fd, ioctlReadTermios, uintptr(unsafe.Pointer(&termios)), 0, 0, 0)
 	return err == 0
 }
